% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno_protein}
\alias{anno_protein}
\title{Protein annotation data}
\format{
## `anno_protein`
A data frame with 2242 rows and 3 columns:
\describe{
  \item{Chr}{Chromosome}
  \item{Start}{Start position}
  \item{Symbol}{Unique symbol for each protein}
  ...
}
}
\usage{
anno_protein
}
\description{
Annotations of each protein from `protein_data`
}
\keyword{datasets}
