% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno_SNP}
\alias{anno_SNP}
\title{SNP annotation data}
\format{
## `anno_SNP`
A data frame with 344388 rows and 3 columns:
\describe{
  \item{CHROM}{Chromosome}
  \item{POS}{Position}
  \item{ID}{Unique identifier for each SNP}
  ...
}
}
\usage{
anno_SNP
}
\description{
Annotations of each SNP from `SNP_data`
}
\keyword{datasets}
