% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_prop.R
\name{cross_prop}
\alias{cross_prop}
\title{Estimation of cellular composition using cross-source information}
\usage{
cross_prop(se, ini_prop, mrk_prot)
}
\arguments{
\item{se}{A `SummarizedExperiment` object with bulk protein expression data frame contained in `counts` slot.}

\item{ini_prop}{Initial cell proportion generated from a different source (e.g. gene expression).
It should be a data frame with rows as matched samples as in protein expression data, and columns as cell types.
The sum of cell proportions for each samples should be 1.}

\item{mrk_prot}{Marker proteins used for cell proportion estimates.}
}
\value{
A `SummarizedExperiment`. The cross-source cell-type proportion estimates for each sample will be stored as an element (`cross_prop`) in `metadata` slot.
}
\description{
This function returns a `SummarizedExperiment` object including cross-source cell-type proportion estimates for each sample.
}
\details{
This is a function developed to implement cell-type proportion deconvolution using initial proportion generated from a different source.
}
