% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csDE.R
\name{csDE}
\alias{csDE}
\title{Cell-type-specific differential expression (csDE)}
\usage{
csDE(se, BPPARAM = bpparam())
}
\arguments{
\item{se}{A `SummarizedExperiment` object with bulk protein/gene expression contained in `counts` slot.
The information from genetic variants should be stored in a P (the number of SNP) by N (the number of samples, should match the sample in `counts` slot) matrix contained as an element (`SNP_data`) in `metadata` slot.
Each matrix entry corresponds to the genotype group indicator (0, 1 or 2) for a sample at a genetic location.
The annotations of these SNP should be stored as an element (`anno_SNP`) in `metadata` slot.
It should include at least the following columns: "CHROM" (which chromosome the SNP is on),
"POS" (position of that SNP) and "ID" (a unique identifier for each SNP, usually a combination of chromosome and its position).
The information on cellular composition is required and stored as `prop` in `metadata` slot.
It is an N (the number of samples, should match the sample in `counts` slot) by K (the number of cell types) matrix.
This can be obtained by running `deconv()` before any filtering steps, or use any source data directly.}
}
\value{
A `SummarizedExperiment`. The csDE results will be stored as an element (`TOAST_output`) in `metadata` slot.
}
\description{
This function returns a `SummarizedExperiment` object including csDE proteins based on samples' genotype.
}
\details{
This is a function developed to implement cell-type-specific differential expression using either `TOAST`.
}
