% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconv}
\alias{deconv}
\title{Estimation of cellular composition in high-throughput data from heterogeneous tissues}
\usage{
deconv(se, method = c("cibersort", "nnls"))
}
\arguments{
\item{se}{A `SummarizedExperiment` object.}

\item{method}{A character string denotes which deconvolution method to use. In thi current version, only `CIBERSORT` or `nnls` is supported.}
}
\value{
A `SummarizedExperiment` with the cell-type proportion estimates for each sample stored in metadata slot.
}
\description{
This function returns a `SummarizedExperiment` object including cell-type proportion estimates for each sample.
}
\details{
This is a function developed to implement cell-type proportion deconvolution using either `CIBERSORT` or `nnls`.
}
\examples{

se <- SummarizedExperiment(assays = list(counts = pQTL::protein_data), rowData = pQTL::anno_protein)
se <- deconv(se, "cibersort")


}
