% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Estimation of cellular composition in high-throughput data from heterogeneous tissues}
\usage{
filter(
  se,
  target_protein = NULL,
  target_SNP = NULL,
  filter_method = c("allele", "distance", "null"),
  filter_allele = 0.25
)
}
\arguments{
\item{se}{A `SummarizedExperiment` object with bulk protein/gene expression contained in `counts` slot.
Annotations on each row (protein or gene) should be stored in rowData() with protein or gene symbol as row names
The first column should be a character vector indicating which chromosome each protein or gene is on.
A "Start" column with numeric values indicating the start position on that chromosome and
a "Symbol" column as a unique name for each protein or gene are also required.
The information from genetic variants should be stored in a P (the number of SNP) by N (the number of samples, should match the sample in `counts` slot) matrix contained as an element (`SNP_data`) in `metadata` slot.
Each matrix entry corresponds to the genotype group indicator (0, 1 or 2) for a sample at a genetic location.
The annotations of these SNP should be stored as an element (`anno_SNP`) in `metadata` slot.
It should include at least the following columns: "CHROM" (which chromosome the SNP is on),
"POS" (position of that SNP) and "ID" (a unique identifier for each SNP, usually a combination of chromosome and its position).}

\item{target_protein}{A character vector contains proteins/genes names that will be used for downstream analysis.
By default, all proteins/genes in `counts` slot will be used.}

\item{target_SNP}{A character vector contains SNP IDs that will be used for downstream analysis.
If not provided, all SNPs will be used for further filtering.}

\item{filter_method}{A character string denotes which filtering method will be used to filter out unrelated SNPs.
If "allele", then the minor allele frequency below argument `filter_allele` will be filtered out.
If "distance", then only cis-acting SNPs for each protein (defined as SNPs on the same chromosome and within 1M base pair (bp) range of that protein) will be included for downstream analysis.
if "null", then the same SNPs will be used for each protein.}

\item{filter_allele}{A numeric value denotes the threshold for minor allele frequency.}
}
\value{
A `SummarizedExperiment`. The results after filtering will be stored as an element (`choose_SNP_list`) in `metadata` slot.
`choose_SNP_list` is a list with the length of the number of proteins for downstream analysis.
Each element stores the index of SNPs to be tested for corresponding protein.
The proteins with no SNPs correspond to it will be removed from the returned list.
}
\description{
This function returns a `SummarizedExperiment` object including cell-type proportion estimates for each sample.
}
\details{
This is a function developed to implement cell-type proportion deconvolution using either `CIBERSORT` or `nnls`.
}
\examples{

se <- SummarizedExperiment(assays = list(counts = pQTL::protein_data), rowData = pQTL::anno_protein)
metadata(se) <- list(SNP_data = pQTL::SNP_data, anno_SNP = pQTL::anno_SNP)
se <- filter(se, filter_method = c("allele", "distance"), filter_allele = 0.25)


}
