% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop_allele}
\alias{prop_allele}
\title{Allele proportion}
\usage{
prop_allele(x)
}
\arguments{
\item{x}{A vector consists of 0, 1 or 2 as genotype group indicator in a genetic location in the genome.}
}
\value{
A numeric value for allele frequency.
}
\description{
This function calculates the allele frequency.
}
