% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop_geno}
\alias{prop_geno}
\title{Genotype proportion}
\usage{
prop_geno(x)
}
\arguments{
\item{x}{A vector consists of 0, 1 or 2 as genotype group indicator in a genetic location in the genome.}
}
\value{
A numeric value for genotype proportion.
}
\description{
This function calculates the genotype proportion in the smallest genotype group.
}
