% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein_data}
\alias{protein_data}
\title{Example protein expression data}
\format{
## `protein_data`
A data frame with 2242 rows (protein) and 20 columns (sample).
}
\usage{
protein_data
}
\description{
A subset of proteomic data PsychENCODE
}
\keyword{datasets}
