% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_data}
\alias{ref_data}
\title{SNP annotation data}
\format{
## `ref_data`
A data frame with 2242 rows (proteins) and 4 columns (cell types).
}
\usage{
ref_data
}
\description{
A signature matrix which serves as a reference of known cellular signatures.
}
\keyword{datasets}
