% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SNP_data}
\alias{SNP_data}
\title{Example SNP data}
\format{
## `SNP_data`
A sparse matrix with 344388 rows (SNP) from one chromosome and 100 columns (sample, should match the sample in `protein_data`).
Each matrix entry corresponds to the genotype group indicator (0, 1 or 2) for a sample at a genetic location.
}
\usage{
SNP_data
}
\description{
A data set stores the information of genetic variants at each location.
}
\keyword{datasets}
